Separate key store tests
